+++ packages/admin-ui/ui/src/components/forms/product/organize-form/use-organize-data.tsx
@@ -3,13 +3,14 @@ import { useMemo } from "react"
 import {
   useAdminCollections,
   useAdminProductCategories,
-  useAdminProductTypes
+  useAdminProductTypes,
 } from "medusa-react"
 
 import { NestedMultiselectOption } from "../../../../domain/categories/components/multiselect"
 import { transformCategoryToNestedFormOptions } from "../../../../domain/categories/utils/transform-response"
 import {
-  FeatureFlag, useFeatureFlag
+  FeatureFlag,
+  useFeatureFlag,
 } from "../../../../providers/feature-flag-provider"
 
 const useOrganizeData = () => {
@@ -19,12 +20,15 @@ const useOrganizeData = () => {
     refetchOnWindowFocus: true,
   })
   const { collections } = useAdminCollections()
-  const { product_categories: categories = [] } =
-    isFeatureEnabled(FeatureFlag.PRODUCT_CATEGORIES) &&
-    useAdminProductCategories({
+  const { product_categories: categories = [] } = useAdminProductCategories(
+    {
       parent_category_id: "null",
       include_descendants_tree: true,
-    })
+    },
+    {
+      enabled: isFeatureEnabled(FeatureFlag.PRODUCT_CATEGORIES),
+    }
+  )
 
   const productTypeOptions = useMemo(() => {
     return (