+++ packages/admin-ui/ui/src/components/forms/product/variant-form/edit-flow-variant-form/index.tsx
@@ -2,7 +2,7 @@ import { useFieldArray, UseFormReturn } from "react-hook-form"
 import CustomsForm, { CustomsFormType } from "../../customs-form"
 import DimensionsForm, { DimensionsFormType } from "../../dimensions-form"
 import VariantGeneralForm, {
-  VariantGeneralFormType
+  VariantGeneralFormType,
 } from "../variant-general-form"
 import VariantStockForm, { VariantStockFormType } from "../variant-stock-form"
 
@@ -11,6 +11,7 @@ import { nestedForm } from "../../../../../utils/nested-form"
 import IconTooltip from "../../../../molecules/icon-tooltip"
 import InputField from "../../../../molecules/input"
 import Accordion from "../../../../organisms/accordion"
+import MetadataForm, { MetadataFormType } from "../../../general/metadata-form"
 import { PricesFormType } from "../../../general/prices-form"
 import VariantPricesForm from "../variant-prices-form"
 
@@ -29,6 +30,7 @@ export type EditFlowVariantFormType = {
   }[]
   customs: CustomsFormType
   dimensions: DimensionsFormType
+  metadata: MetadataFormType
 }
 
 type Props = {
@@ -115,13 +117,22 @@ const EditFlowVariantForm = ({ form, isEdit }: Props) => {
           </p>
           <DimensionsForm form={nestedForm(form, "dimensions")} />
         </div>
-        <div className="mt-xlarge">
-          <h3 className="inter-base-semibold mb-2xsmall">Customs</h3>
-          <p className="inter-base-regular text-grey-50 mb-large">
-            Configure if you are shipping internationally.
-          </p>
-          <CustomsForm form={nestedForm(form, "customs")} />
-        </div>
+        {showStockAndInventory && (
+          <div className="mt-xlarge">
+            <h3 className="inter-base-semibold mb-2xsmall">Customs</h3>
+            <p className="inter-base-regular text-grey-50 mb-large">
+              Configure if you are shipping internationally.
+            </p>
+            <CustomsForm form={nestedForm(form, "customs")} />
+          </div>
+        )}
+      </Accordion.Item>
+      <Accordion.Item title="Metadata" value="metadata">
+        <p className="inter-base-regular text-grey-50 mb-base">
+          Metadata can be used to store additional information about the
+          variant.
+        </p>
+        <MetadataForm form={nestedForm(form, "metadata")} />
       </Accordion.Item>
     </Accordion>
   )