+++ packages/admin-ui/ui/src/domain/customers/groups/details.tsx
@@ -1,6 +1,5 @@
-import { useContext, useEffect, useState } from "react"
-import { difference } from "lodash"
 import { CustomerGroup } from "@medusajs/medusa"
+import { difference } from "lodash"
 import {
   useAdminAddCustomersToCustomerGroup,
   useAdminCustomerGroup,
@@ -8,20 +7,21 @@ import {
   useAdminDeleteCustomerGroup,
   useAdminRemoveCustomersFromCustomerGroup,
 } from "medusa-react"
+import { useEffect, useState } from "react"
 
-import BodyCard from "../../../components/organisms/body-card"
-import EditIcon from "../../../components/fundamentals/icons/edit-icon"
-import TrashIcon from "../../../components/fundamentals/icons/trash-icon"
-import PlusIcon from "../../../components/fundamentals/icons/plus-icon"
-import EditCustomersTable from "../../../components/templates/customer-group-table/edit-customers-table"
-import CustomersListTable from "../../../components/templates/customer-group-table/customers-list-table"
-import CustomerGroupContext, {
-  CustomerGroupContextContainer,
-} from "./context/customer-group-context"
-import useQueryFilters from "../../../hooks/use-query-filters"
-import DeletePrompt from "../../../components/organisms/delete-prompt"
 import { useNavigate, useParams } from "react-router-dom"
 import BackButton from "../../../components/atoms/back-button"
+import EditIcon from "../../../components/fundamentals/icons/edit-icon"
+import PlusIcon from "../../../components/fundamentals/icons/plus-icon"
+import TrashIcon from "../../../components/fundamentals/icons/trash-icon"
+import { ActionType } from "../../../components/molecules/actionables"
+import BodyCard from "../../../components/organisms/body-card"
+import DeletePrompt from "../../../components/organisms/delete-prompt"
+import CustomersListTable from "../../../components/templates/customer-group-table/customers-list-table"
+import EditCustomersTable from "../../../components/templates/customer-group-table/edit-customers-table"
+import useQueryFilters from "../../../hooks/use-query-filters"
+import useToggleState from "../../../hooks/use-toggle-state"
+import CustomerGroupModal from "./customer-group-modal"
 
 /**
  * Default filtering config for querying customer group customers list endpoint.
@@ -164,7 +164,6 @@ type CustomerGroupDetailsHeaderProps = {
  * Customers groups details page header.
  */
 function CustomerGroupDetailsHeader(props: CustomerGroupDetailsHeaderProps) {
-  const { showModal } = useContext(CustomerGroupContext)
   const [showDeleteConfirmation, setShowDeleteConfirmation] = useState(false)
 
   const navigate = useNavigate()
@@ -173,10 +172,12 @@ function CustomerGroupDetailsHeader(props: CustomerGroupDetailsHeaderProps) {
     props.customerGroup.id
   )
 
-  const actions = [
+  const { state, close, open } = useToggleState()
+
+  const actions: ActionType[] = [
     {
       label: "Edit",
-      onClick: showModal,
+      onClick: open,
       icon: <EditIcon size={20} />,
     },
     {
@@ -214,6 +215,11 @@ function CustomerGroupDetailsHeader(props: CustomerGroupDetailsHeaderProps) {
           text="Are you sure you want to delete this customer group?"
         />
       )}
+      <CustomerGroupModal
+        open={state}
+        onClose={close}
+        customerGroup={props.customerGroup}
+      />
     </>
   )
 }
@@ -231,17 +237,15 @@ function CustomerGroupDetails() {
   }
 
   return (
-    <CustomerGroupContextContainer group={customer_group}>
-      <div className="-mt-4 pb-4">
-        <BackButton
-          path="/a/customers/groups"
-          label="Back to customer groups"
-          className="mb-4"
-        />
-        <CustomerGroupDetailsHeader customerGroup={customer_group} />
-        <CustomerGroupCustomersList group={customer_group} />
-      </div>
-    </CustomerGroupContextContainer>
+    <div className="-mt-4 pb-4">
+      <BackButton
+        path="/a/customers/groups"
+        label="Back to customer groups"
+        className="mb-4"
+      />
+      <CustomerGroupDetailsHeader customerGroup={customer_group} />
+      <CustomerGroupCustomersList group={customer_group} />
+    </div>
   )
 }