+++ packages/medusa-payment-stripe/src/api/utils/utils.ts
@@ -1,13 +1,13 @@
-import { AwilixContainer } from "awilix"
-import Stripe from "stripe"
 import {
   AbstractCartCompletionStrategy,
   CartService,
   IdempotencyKeyService,
   PostgresError,
 } from "@medusajs/medusa"
-import { EOL } from "os"
+import { AwilixContainer } from "awilix"
 import { MedusaError } from "medusa-core-utils"
+import { EOL } from "os"
+import Stripe from "stripe"
 
 const PAYMENT_PROVIDER_KEY = "pp_stripe"
 
@@ -251,8 +251,8 @@ async function completeCartIfNecessary({
     if (response_code !== 200) {
       throw new MedusaError(
         MedusaError.Types.UNEXPECTED_STATE,
-        response_body["message"],
-        response_body["code"].toString()
+        response_body["message"] as string,
+        response_body["code"] as string
       )
     }
   }