+++ packages/medusa/src/api/routes/admin/orders/fulfill-swap.ts
@@ -6,8 +6,9 @@ import {
 } from "../../../../services"
 
 import { EntityManager } from "typeorm"
-import { validator } from "../../../../utils/validator"
 import { FindParams } from "../../../../types/common"
+import { cleanResponseData } from "../../../../utils/clean-response-data"
+import { validator } from "../../../../utils/validator"
 import { updateInventoryAndReservations } from "./create-fulfillment"
 
 /**
@@ -134,7 +135,7 @@ export default async (req, res) => {
     includes: req.includes,
   })
 
-  res.status(200).json({ order })
+  res.status(200).json({ order: cleanResponseData(order, []) })
 }
 
 /**